Move debouncer function into the only file that uses it
