+++ ui/src/app/utils/index.tsx
@@ -374,26 +374,6 @@ export function hasNewValidProps(currProps, prevProps, fieldsToCompare) {
   return false;
 }
 
-// Returns a function which will execute `action` at most once every `sensitivityMs` milliseconds
-// if the returned function has been invoked within the last `sensitivityMs` milliseconds
-// Example : debouncing user activity events to change rate of invocation from 1000/s to 1/s
-export function debouncer(action, sensitivityMs) {
-  let t = Date.now();
-
-  const timer = global.setInterval(() => {
-    if (Date.now() - t < sensitivityMs) {
-      action();
-    }
-  }, sensitivityMs);
-
-  return {
-    invoke: () => {
-      t = Date.now();
-    },
-    getTimer: () => timer,
-  };
-}
-
 // Starts a timer which will invoke `f` after `timeoutInSeconds` has passed
 // Calling the returned function will reset the timer to zero
 // Example : Call a logout function after 30 seconds of the returned function not being invoked
